fix: don't throw an exception if save recent goes out of sync

+++ packages/compass-connect/src/stores/index.js
@@ -931,6 +931,11 @@ const Store = Reflux.createStore({
         (item) => item._id === recents[9]
       );
 
+      // somehow this can happen
+      if (!toDestroy) {
+        return;
+      }
+
       toDestroy.destroy({
         success: () => {
           this.state.fetchedConnections.remove(toDestroy._id);
